test/mochitest/test_cleanup_imagingtools.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
</div>
<pre id="test">
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript">

  "use strict";

  var realSendSyncMessage;

  function test_setup() {
    return new Promise((resolve, reject) => {
      is(typeof _isRisky, "boolean", "has risky boolean");
      _isRisky = false;
      is(typeof cpmm, "object", "has cpmm object");
      is(typeof cpmm.sendSyncMessage, "function", "has cpmm.sendSyncMessage");
      realSendSyncMessage = cpmm.sendSyncMessage;
      resolve();
    });
  }

  function test_unload() {
    return new Promise((resolve, reject) => {
      cpmm.sendSyncMessage = (msg, data) => {
        is(msg, "B2GInstaller:MainProcess:Cleanup", "beforeunload sends cleanup");
        cpmm.sendSyncMessage = realSendSyncMessage;
        resolve();
      };
      window.dispatchEvent(new Event("beforeunload"));
    });
  }

  function runTest() {
    loadAboutJs()
      .then(() => test_setup())
      .then(() => test_unload())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
